pyproject.toml
[tool.poetry]
name = "czech-holidays"
version = "1.0.0"
description = "Python package with Czech public holidays"
authors = ["Honza Javorek <mail@honzajavorek.cz>"]
license = "MIT"
readme = "README.rst"
repository = "https://github.com/honzajavorek/czech-holidays"

[tool.poetry.dependencies]
python = ">=3.7"
python-dateutil = "*"

[tool.poetry.dev-dependencies]
pytest = "*"

[tool.pytest.ini_options]
python_files = "test_*.py"
testpaths = "."
addopts = "--ff --doctest-glob=README.rst"
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"
